# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Management API",
        "description": "API documentation for managing users including CRUD operations and currency information retrieval.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/countries-with-currencies": {
            "get": {
                "tags": [
                    "Country"
                ],
                "summary": "Get all countries with currencies",
                "operationId": "e3c3936124231ae9fb30c523026d9183",
                "responses": {
                    "200": {
                        "description": "A list of countries with their currencies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code": {
                                                "type": "string",
                                                "example": "US"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "United States"
                                            },
                                            "currencies": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "code": {
                                                            "type": "string",
                                                            "example": "USD"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "United States Dollar"
                                                        },
                                                        "symbol": {
                                                            "type": "string",
                                                            "example": "$"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User - Get all users with filtering and sorting",
                "description": "Retrieve a list of all users, with optional filters for country and currency, and sorting options.",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "parameters": [
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Filter by country name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "description": "Filter by currency code",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by (e.g., name, email, country, or currency)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "email",
                                "country",
                                "currency"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (ascending or descending)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "mohammad"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "mohammad.hamzei1990@gmail.com"
                                            },
                                            "country": {
                                                "type": "string",
                                                "example": "South Georgia"
                                            },
                                            "currency": {
                                                "type": "string",
                                                "example": "SHP"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User - Create a new user",
                "description": "Add a new user with a specific country and currency",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "country_id",
                                    "currency_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "mohammad"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mohammad.hamzei1990@gmail.com"
                                    },
                                    "country_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "currency_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "400": {
                        "description": "Validation error or currency does not belong to the country"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User - Get user by ID",
                "description": "Retrieve user information by ID",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "User - Update user information",
                "description": "Update user data by ID",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "updated@example.com"
                                    },
                                    "country_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "currency_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "User - Delete user",
                "description": "Delete a user by ID",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Country",
            "description": "API Endpoints of Countries"
        },
        {
            "name": "User",
            "description": "API Endpoints of Users"
        }
    ]
}